feat: hides astar implementation details
